backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
numpy==1.26.4
silero-vad==5.1.2
qdrant-client==1.12.1
sentence-transformers==3.3.1
asyncpg==0.29.0
python-escpos==3.0
groq>=0.14.0
openai>=1.0.0
google-generativeai>=0.8.0
httpx>=0.27.0,<0.28.0
aiohttp==3.11.10
sarvamai>=0.1.21
